backend/requirements_simple.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
# Simplified scientific dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
# AI Dependencies
openai>=1.0.0
scikit-learn>=1.0.0
joblib>=1.0.0